Added options.type and options.format
options.type is a string that defines how the result array will be populated. ( I implemented the "PAIRS" type ).
options.format is a string that defines how words will be formatted. (I implemented the "PASCALCASE" type).
if type and format are not defined or are different than any case in the switch statements, than it all works as default.

I implemented this because I published an article on medium, I made a React Native version of the Flutter's "Write your first Flutter app". ( you can find it here: https://medium.com/@mattveraldi/flutter-write-your-first-flutter-app-in-react-native-1e468ab8655b ).
I used a function on top of random-words to do the same thing that english_words Flutter's package does. So I thought to make a pull request so anyone can use it.